usb: Fix CHERI issues for drivers/usb/core/
- Add neccessary conversions.
- Do not store a pointer unaligned for CHERI.

+++ drivers/usb/core/buffer.c
@@ -155,7 +155,7 @@ void hcd_buffer_free(
 		return;
 
 	if (hcd->localmem_pool) {
-		gen_pool_free(hcd->localmem_pool, (unsigned long)addr, size);
+		gen_pool_free(hcd->localmem_pool, (uintptr_t)addr, size);
 		return;
 	}
 
@@ -202,7 +202,7 @@ void hcd_buffer_free_pages(struct usb_hcd *hcd,
 
 	if (hcd->localmem_pool) {
 		gen_pool_free(hcd->localmem_pool,
-				(unsigned long)addr, size);
+				(uintptr_t)addr, size);
 		return;
 	}
